+++ spring-boot-email-core/src/main/java/it/ozimov/springboot/templating/mail/model/defaultimpl/DefaultEmailAttachment.java
@@ -39,7 +39,6 @@ import static java.util.Optional.ofNullable;
 /**
  * Plain text email.
  *
- * @author rtrunfio
  */
 @NoArgsConstructor
 @RequiredArgsConstructor